+++ src/practice_fixes/play_layer.cpp
@@ -26,15 +26,11 @@ class $modify(GJBaseGameLayer) {
 };
 
 class $modify(CheckpointObject) {
-  #ifdef GEODE_IS_WINDOWS
+
+  // CheckpointObject::init is available on every platform since 2.2081
   bool init() {
     bool ret = CheckpointObject::init();
     CheckpointObject* cp = this;
-  #else
-  static CheckpointObject* create() {
-    CheckpointObject* ret = CheckpointObject::create();
-    CheckpointObject* cp = ret;
-  #endif
 
     if (!cp) return ret;
 
@@ -48,11 +44,7 @@ class $modify(CheckpointObject) {
       Global::getCurrentFrame(),
       p1Data,
       p2Data,
-      #ifdef GEODE_IS_WINDOWS
-      *(uintptr_t*)((char*)geode::base::get() + seedAddr),
-      #else
-      0,
-      #endif
+      static_cast<uintptr_t>(GameToolbox::getfast_srand()),
       Global::get().previousFrame
     };
 
@@ -117,14 +109,8 @@ class $modify(PlayLayer) {
     g.ignoreJumpButton = frame + 1;
     g.previousFrame = g.checkpoints[cp].previousFrame;
 
-    #ifdef GEODE_IS_WINDOWS
-
-    if (g.seedEnabled) {
-      uintptr_t seed = g.checkpoints[cp].seed;
-      *(uintptr_t*)((char*)geode::base::get() + seedAddr) = seed;
-    }
-
-    #endif
+    if (g.seedEnabled)
+      GameToolbox::fast_srand(static_cast<uint64_t>(g.checkpoints[cp].seed));
 
     if (g.state == state::recording)
       InputPracticeFixes::applyFixes(this, p1Data, p2Data, frame);