+++ src/ui/macro_editor.hpp
@@ -11,7 +11,7 @@ struct InputText {
     std::string action;
 };
 
-class MacroEditLayer : public geode::Popup<>, public TextInputDelegate {
+class MacroEditLayer : public xdbot::Popup<>, public TextInputDelegate {
 
 private: