+++ src/ui/macro_info_layer.hpp
@@ -2,7 +2,7 @@
 #include "../includes.hpp"
 #include <Geode/ui/MDTextArea.hpp>
 
-class MacroInfoLayer : public geode::Popup<> {
+class MacroInfoLayer : public xdbot::Popup<> {
 
 public: