+++ src/ui/record_layer.hpp
@@ -31,7 +31,7 @@ const float ySettingPositions[6]{
 	76.5f, 47.5f, 18.5f, -11.5f, -40.5f, -69.5f
 };
 
-class RecordLayer : public geode::Popup<>, public TextInputDelegate {
+class RecordLayer : public xdbot::Popup<>, public TextInputDelegate {
 public:
 	CCMenuItemToggler* recording = nullptr;
 	CCMenuItemToggler* playing = nullptr;