+++ src/ui/render_settings_layer.hpp
@@ -2,7 +2,7 @@
 
 #include "../includes.hpp"
 
-class RenderSettingsLayer : public geode::Popup<>, public TextInputDelegate {
+class RenderSettingsLayer : public xdbot::Popup<>, public TextInputDelegate {
 	
 public: